% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R, R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)

fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{as a variable that represent a state.}

\item{year}{as a variable to tranform it in an integer value.}
}
\value{
if number of a state is unique and it is contained in the variable STATE of the data
it will make a data frame with this filter, with conditionals to transform NAs in the
variables LONGITUD AND LATITUDE and print a map with this location. Otherwise print a
message "no accidents to plot" and return an invisible object.

if number of a state is unique and it is contained in the variable STATE of the data
it will make a data frame with this filter, with conditionals to transform NAs in the
variables LONGITUD AND LATITUDE and print a map with this location. Otherwise print a
message "no accidents to plot" and return an invisible object.
}
\description{
The function \code{fars_map_state} transform the principal database and add
conditionals for some variables. Plot a map with a specific lat and long.

The function \code{fars_map_state} transform the principal database and add
conditionals for some variables. Plot a map with a specific lat and long.
}
\details{
you need to install the package "map" and specify a number of a state.

you need to install the package "map" and specify a number of a state.
}
\examples{
fars_map_state(19, 2014)
fars_map_state(19, 2014)
}
