% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R, R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)

fars_read(filename)
}
\arguments{
\item{filename}{to enter a database with format csv.}
}
\value{
if file exists, this function read the file and return a database as a data frame. If the extension
is diferent to csv, it can not read the file.

if file exists, this function read the file and return a database as a data frame. If the extension
is diferent to csv, it can not read the file.
}
\description{
The function \code{fars_read} read a csv file if it exists and forwards the argument a data frame.

The function \code{fars_read} read a csv file if it exists and forwards the argument a data frame.
}
\details{
There functions used a database to create maps according a specific location.

you need install packages like dplyr and readr before this or it may result in an error

you need install packages like dplyr and readr before this or it may result in an error
}
\examples{
fars_read(filename = accident_2013.csv.bz2)
fars_read(filename = accident_2013.csv.bz2)
}
