% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R, R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)

fars_read_years(years)
}
\arguments{
\item{years}{as a variable to tranform it in an integer value accross \code{make_filename}. It will be used
when it generate the name of a file with the function \code{fars_read}.}
}
\value{
if the year is in the set: 2013, 2014 or 2015, it will transmute a database in a new data frame depending
on a specific month. Otherwise it will print a warning.

if the year is in the set: 2013, 2014 or 2015, it will transmute a database in a new data frame depending
on a specific month. Otherwise it will print a warning.
}
\description{
The function \code{fars_read_years} save the name of a specific data base according to a year,
read the ddatabase and transmute it drops existing variables as year.

The function \code{fars_read_years} save the name of a specific data base according to a year,
read the ddatabase and transmute it drops existing variables as year.
}
\details{
you need enter a number as a year contained in the set: 2013, 2014 or 2015 or it will return a warning
as a message.

you need enter a number as a year contained in the set: 2013, 2014 or 2015 or it will return a warning
as a message.
}
\examples{
\dontrun{fars_read_years(2013)}
fars_read_years(2013)
}
\seealso{
\code{\link{fars_read}}

\code{\link{make_filename}}
}
