% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{as a variable to read the other functions and transmute the data frame.}
}
\value{
a data frame by group of year and month, and summarize by count. It will print the head of the database.
}
\description{
The function \code{fars_summarize_years} transmute the data frame by group, summarize and
spread a key-value pair across the variables year and n.
}
\details{
you need install the library tidyr and conserve the format of the variables.
}
\examples{
fars_summarize_years(2014)
}
