% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{as a variable to tranform it in an integer value.}
}
\value{
a character vector containing a formatted combination of text and variable value.
}
\description{
The function \code{make_file} transform a variable as a integer and print a character vector
containing a formatted combination of text and variable value.
}
\details{
you need enter a number or it return a NA.
}
\examples{
make_filename(2014)
}
